# swagger/api.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Atomica Blogspot"
basePath: "/v1"
...
schemes:
  - "https"
  - "http"
paths:
...
  /post/{postId}:
    get:
      tags:
        - "post"
      summary: "Get post by ID"
      description: "Returns a single post by its id"
      operationId: "getPostById"
      produces:
        - "application/json"
      parameters:
        - name: "postId"
          in: "path"
          description: "ID of post to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Post"
        "400":
          $ref: '#/responses/invalid-request'
        "404":
          $ref: '#/responses/not-found'
      security: []
    put:
      tags:
        - "post"
      summary: "Update an existing post"
      description: ""
      operationId: "updatePost"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "postId"
          in: "path"
          description: "ID of post to return"
          required: true
          type: "integer"
          format: "int64"
        - in: body
          name: post
          schema:
            $ref: '#/definitions/Post'
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Post not found"
        "405":
          description: "Validation exception"
    delete:
      tags:
        - "post"
      summary: "Soft Delete a post"
      description: ""
      operationId: "deletePost"
      produces:
        - "application/json"
      parameters:
        - name: "postId"
          in: "path"
          description: "Post id to delete"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: success
        "400":
          $ref: '#/responses/invalid-request'
        "404":
          $ref: '#/responses/not-found'
      security: []
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Post:
    type: "object"
    required:
      - "title"
      - "photoUrls"
      - "content"
    properties:
      id:
        type: "integer"
        format: "int64"
      title:
        type: "string"
        example: "Liverpool achieved astonishing win"
      content:
        type: "string"
        example: "Liverpool acheived astonishing win"
      photoUrls:
        type: "array"
        items:
          type: "string"
      tags:
        type: "array"
        items:
          type: "string"
      status:
        type: "string"
        description: "post status in the blog"
        enum:
          - "published"
          - "deleted"
...
  error:
    type: object
    title: Error Response
    description: Standard error format
    properties:
      code:
        type: string
      message:
        type: string
responses:
  invalid-request:
    description: Invalid request
    schema:
      $ref: '#/definitions/error'
  not-found:
    description: Not found.
    schema:
      $ref: "#/definitions/error"